Refactored FileDescriptionHeader to traceability
